v2/bird-node/requirements.txt
numpy==2.2.4
python-dotenv==1.0.1
sounddevice==0.5.1
soundfile==0.13.1
